+++ ReleaseNotes.md
@@ -2,6 +2,13 @@
 
 ## 1.35
 
+### admin-users
+
+Fixed user search when one or many of these user data values  is not defined: 
+- user name
+- firstname
+- last name
+
 ### mapwfs2_ol3
 
 WFSRefreshManualLoadLayersEvent is now included in ol3 version as well and changing the size of the map no longer results in JS-error.